Tests/LibWeb/Text/expected/canvas/drawimage-webgl-canvas-source.txt

B sampled the WebGL frame at drawImage time (red): true
